services/embedder/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.16.1
fsspec==2025.3.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
hf-xet==1.2.0
httptools==0.6.4
huggingface-hub==0.25.2
idna==3.11
jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.9.1
numpy==1.24.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==10.4.0
protobuf==5.29.5
pydantic==2.5.0
pydantic-core==2.14.1
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch --index-url https://download.pytorch.org/whl/cpu
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
flask
